Fix bug where setup trys to unlink a non existant file.

+++ packages/react-atlas/src/setup.js
@@ -236,7 +236,9 @@ function createComponentDirectories() {
     component = component[0].toUpperCase() + component.slice(1);
 
     if(fs.existsSync(oldconfigPath + component + '/README.md')) {
-      fs.unlinkSync(componentDirconfigPath + '/' + component + '/README.md');
+      if(fs.existsSync(componentDirconfigPath + '/' + component + '/README.md')) {
+        fs.unlinkSync(componentDirconfigPath + '/' + component + '/README.md');
+      }
       console.log("Linking: %s -> %s", oldconfigPath + component + '/README.md', componentDirconfigPath + '/' + component + '/README.md');
       fs.linkSync(oldconfigPath + component + '/README.md', componentDirconfigPath + '/' + component + '/README.md');
     }
